fix(B-464): split admin stats model and schema
